requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Audio processing
pydub>=0.25.1

# Data models
pydantic>=2.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Utilities
aiofiles>=23.2.1

